+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat: Add support for Dynamic Sampling (#661)
 - Add support for tracing of the Symfony HTTP client requests (#606)
 - Support logging the impersonator user, if any (#647)