+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat: Add support for Dynamic Sampling (#661)
 - Add support for tracing of the Symfony HTTP client requests (#606)
 - Support logging the impersonator user, if any (#647)
+- ref: Use constant for the SDK version (#662)
 
 ## 4.3.1 (2022-10-10)